src/requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
lxml==4.3.4
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.1.0
pyaes==1.6.1
pyasn1==0.4.5
requests==2.22.0
rsa==4.0
Telethon==1.8.0
urllib3==1.25.3
